Move karma.config.shared.js -> karma.shared.js
